pyproject.toml
[project]
name = "vacation-planner"
version = "1.0.0"
description = "AI Vacation Planner PoC"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.37.0",
    "pandas>=2.2.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "langgraph>=0.2.24",
    "langchain-core>=0.3.12",
    "groq>=0.12.0",
    "python-dateutil>=2.9.0.post0",
]

[tool.uv]